Accessibility improvements (#12)
Fixes neevaco/neeva-ios#53. I tested everything using VoiceOver. The only thing that’s not working right now is the thumbnail picker. It has a very strange animation not present when opening the screen without VoiceOver, and it’s difficult to navigate between the thumbnail choices. I could not figure out why this was happening or how to correct it.