Update urls used for address bar search/suggest requests (#33)
Two changes here:

(1) Change src param on search and suggest URLs to nvobar (stands for: Neeva Omnibar). This is the same value we use for the default search provider URL set from our browser extension. This value is used for log analysis. I checked with Avi, and he'd prefer to not introduce a new src param value here, and he was ok with either nvobar or opensearch. I feel nvobar makes a bit more sense, but wdyt @zimbabao?

(2) Use the appHost constant to build these URLs, so that the URLs will point at m1, etc. if user changes this in the debug settings.