Make sure SwiftUI components honor system dark mode setting properly (#54)
Implement DarkTheme's userInterfaceStyle property to return .dark overridding
its base class's .light value. Seems like an obvious oversight in the code that
it didn't do this and probably what was originally intended.

Also this change adds code to observe changes to the theme at a few additional
spots matching the approach taken elsewhere. This way if the user changes the
system settings, our UI will update properly.

Another approach would be to set overrideUserInterfaceStyle to .unspecified
when we want to honor system settings, but the approach taken in this PR seems
to be more aligned to the intent of the existing code.

Closes #23