Fix "Add to Space" hang. (#22)
Move .onDisappear handler to the LoadingView from the top-level View so that
we don't try to cancel the request when updating the top-level View to show
the LoadingView.

Since the View transitions asynchronously to show the LoadingView, there is a
small moment in time when the View is not yet showing the LoadingView and could
have a non-nil cancellation object. If the View is torn down at that point, we
could end up leaving the cancellation object running in the background. That
means the network request could still complete. It also means the network
request, if it is super slow, could hang around for a long time. The network
request does keep the AddToSpaceView instance alive by virtue of that being
captured by the result closure. There are still some potential issues with
this to work out, and this PR does not attempt to resolve all of these issues.

Also remove unused isWaiting state variable.

Fixes #15