Inject cookie to self-identify device type as ios-browser (#59)
We need the serving stack to differentiate requests coming from the iOS browser app (so we can do things like hide the search bar in the UI). We've decided to use the "device type" value in the serving context for this purpose. Previously, we were intercepting all GET requests to our domain, adding the X-Device-Type-Header (which is used to populate the serving context in our serving stack), then restarting the request.

An alternative approach taken here is to simply inject a cookie with the same value. We'll also need to update httpd to respect this cookie when populating the serving context.