URL bar should fade to white (not black) when page is scrolled (#156)
The solution is to just not set an alpha on URLBarView when asked to set an
alpha value on subviews. This makes sense since URLBarView is not one of the
subviews. The Neeva menu button needs to have the alpha set on it, so that it
fades away properly.

Fixes #136