Implement more of the incognito UX (#165)
Switches URL bar background to be black when in incognito mode.  As part of
this, cleans up some of the applyTheme and applyUIMode code in URLBarView and
associated classes.

Deletes unnecessary extra shield icons. It turns only need one to serve as a
template that can be tinted to different colors.

Issue #154