Change spaces option from Neeva menu to open the web version of spaces (#168)
The native implementation was not feature complete and would have required more
work to bring it up to parity with what we have on the web. This also makes it
easier to explore spaces from within the browser by leveraging the tab browsing
UI.

This change also fixes a bug with the home option that would cause it to not go
to Neeva home if the user had configured their New Tab Page to be something
other than the default option.

Finally, both Home and Spaces are made to open in a new tab rather than load in
the current tab. This way the current journey is not clobbered when someone
wants to browse spaces. This could result in extra tab clutter, and we should
address that but in a different way.

Fixes #167