Implements half-height modal for add-to-spaces flow (#191)
* Implements half-height modal for add-to-spaces flow

This is based on the OverlayView.swift from neeva-ios-playground, but a lot
changed since that prototype as a result of integrating it with AddToSpaceView.
Some things are simplified from that prototype, such as enabling the list to be
scrollable when in half-height mode. This means the dialog is completely usable
without expanding it to full height.

This change also includes some reworking of AddToSpaceView to fit progress and
error UI into the same ScrollView used to hold the content. That avoids some
flicker when transitioning from loading to showing the actual content. It also
means the error page will fit better in half-height + landscape mode.

Added a tap gesture on the sheet whitespace to help provide a way to dismiss
the virtual keyboard if it is open as a result of the user typing a filter
string. Previously the keyboard would remain open until the dialog closed.

Still to-do is updating metrics and colors to better conform to mocks.

This change also improves the add-to-spaces flow when sharing to the Neeva
from another browser or app. It avoids some redundant Save to Spaces text
and also works better with minimal height since the filter box is included
in the scrollable region.

Removed the .refreshControl View extension applied to the spaces list as it
wasn't actually working. It appears to only work with TableView and not
ScrollView. I think it is okay to not have pull-to-refresh as part of this.

Issue #45

* Ensure backdrop covers all edges