Fix the layout bounds of SearchSuggestionView (#210)
This turned out to be a fairly complicted thing to fix, and the fix is a bit of
a hack. It seems like there might be a SwiftUI bug that makes this hard to fix
in what should be a straightforward manner.

First, it was important to embed the ErrorView in a ScrollView so that it would
not extend beyond the provided bounds and so that it would be usable in
landscape mode. The non-error case already handles this by using a List view
(which is similar to a ScrollView).

Next, it was important to position that ScrollView (and the List) properly. The
existing code tried to do so by setting bottom padding equal to the measured
height of the virtual keyboard and telling SwiftUI to ignore the safe area at
the bottom (i.e., to ignore the keyboard vertical height).

The above is pretty close to being correct except that it doesn't take into
account what happens when the device geometry changes. For example, if the
device is rotated, then the keyboard height needs to be recalculated.

The way to know when the device geometry changes in SwiftUI is to instead just
parameterize the layout by the device geometry. This is where GeometryReader
comes in. It exposes the geometry of the parent view, and we can layout our
views according to the size of the parent.

If SwiftUI sees you not only using a GeometryReader but importantly reading the
GeometryProxy value provided by it, then it will think that part of your view
tree depends on the GeometryProxy values and it will update your view tree when
those values change.

Great, except that the size of the padding doesn't actually depend on the
values returned by GeometryProxy. It is just determined by the height of the
virtual keyboard.

Here's the hack. We need SwiftUI to query the height of the virtual keyboard
when that changes. We could probably set-up a way to do this with Combine
by exposing the keyboard height as a publisher.

Another way is to just have a function that takes as input the GeometryProxy,
reads that but only returns the height of the keyboard. That's sufficient to
convince SwiftUI that it should call that function when needed.

And that's what this change does. There has to be a better way!

(Also, instead of setting bottom padding on the ScrollView and List, this
change just puts both of those in a VStack and places a Spacer below them
that has its height set equivalently to what the bottom padding would be.
This way the value is only set in one place.)

Fixes #207