Use a better approach to sizing the OverlaySheet relative to the keyboard (#215)
Observe keyboard height and expose via a Publisher (using Combine) so that
SwiftUI can observe and respond to it. Previously we were just observing
keyboard visibility state. But we can extend this approach to get its height.

One subtle detail is that we observe keyboardDidShowNotification rather than
keyboardWillShowNotification. That way we start our animation to adjust the
OverlaySheet after the keyboard is done animating. This way it looks like the
OverlaySheet is resized in response to the keyboard and it avoids an issue with
the OverlaySheet appearing to jump into its new position (leaving a visible
gap) before the keyboard has finished sliding up into place.

With the keyboard height, we can position the OverlaySheet exactly above the
keyboard. Previously we were relying on SwiftUI's automatic way of positioning
content above the keyboard when not ignoring the safe area at the bottom. This
is why the previous code would switch in and out of ignoring the bottom safe
area based on whether the keyboard was visible or not. But, this caused a
visible tearing as SwiftUI would adjust the position of the UI too soon. We
didn't have control over this, so going down the path of manually positioning
with keyboard height works better.

Fixes another UI glitch with a gray rectangle appearing to slide down as the
OverlaySheet it sliding up when first opening. This was fixed by delaying
the call to show the OverlaySheet via DispatchQueue.main.async. Not sure I
can fully explain why that is needed, but it does seem to work.

Speeds up animation of OverlaySheet a bit.

Removes tap gesture handling on OverlaySheet used to dismiss keyboard. User can
just tap the "Return" button on the keyboard to dismiss. This tap handler was
interfering with other tap gesture handling.

Issue #126