Enable SharedTests and add to CircleCI (#220)
This test suite is fairly trivial and happens to still work :) It is a good way
to explore getting automated testing on CircleCI.

Ultimately, I think we will want to use separate jobs for the different test
suites. That way they run in parallel. But leaving that for an exercise when we
have multiple test suites!

Issue #5