Enable StorageTests (#228)
Removes sync commands tests as we don't use mozilla's sync system.

Disables some tests that do not function:
- RustLogins because we don't have a backend for login storage.
- TestSwiftData because it seems to hang.
- testMovesDB because it fails.

Adds back some bookmarks strings needed by TestSQLiteHistory. It may be the
case that we can avoid this by deleting some portion of the tests, but I'm
saving that for a future investigation.

Adds back DeferredTestUtils.swift which was deleted for an unexplained reason
by a previous commit. This was needed to compile StorageTests.

Issue #5