Ensure bottom bar isn't left in a partially visible state. (#233)
The TabScrollController was only basing its decision on being done with
animating the top and bottom bar on the position of the top bar. It needs to
also take into account the current position of the bottom bar. Only when both
are fully visible should it set the toolbarState to .visible.

This change also changes where we stop the UIRefreshControl animating on
pull-to-refresh. This is moved up earlier since after the page reload is
started, the progress bar can take over to display load progress. This feels
like an improvement so we don't have as much animating at the same time during
page reload.

Fixes #224