Restore Firefox code for setting background color during tab animations (#241)
This was removed as a side-effect of the mega Neeva Base Tasks PR.  That was a
project to apply code changes made by Jed to an updated version of Firefox.
Unclear why this code was removed.

Restoring this code makes it so that there is no longer a black flash when
opening new tabs to web content (e.g., Spaces). It also makes it so that there
is less of a visible "coming together" of the top of the tab when being
selected from the tab grid. There is still some artifact there, but that can
be resolved as another change.

Issue #163