Remove Sentry dependency. (#250)
This keeps most of the Sentry integration in place and just drops the
third-party component. We might want to bring this back if we decide
to use Sentry. Alternatively, we might want to just hook this up to
our own Client logging system. In that case, all of the existing hooks
in the code will be useful and we can just rename SentryIntegration.

Issue #96