Make OverlaySheetView configurable (#267)
Allow hiding the title and setting a custom background color.

This change also fixes some visual glitches (that horizontal line) when
resizing the sheet by applying the background to the entire sheet instead
of applying it to the individual segments. Also moves the cornerRadius
to that background so that it can apply properly when the title bar is
very short in height.