Improve OverlaySheet animation when keyboard is showing or hiding (#270)
Previously the fixed-height mode animation when the keyboard appears or
disappears was pretty janky. The sheet would resize at a different rate than
the embedded content, making it seem like the two are disconnected.

The solution is to change getSpacerHeight to take keyboardHeight into account.
To do that the outerGeometry, the GeometryReader used to determine spacer
height, needs to be based on ignoring the bottom safe area so that it would
not depend on keyboard height itself.

As a result now, both the spacer above and below the contents of the sheet
would be adjusted together and in a consistent way. That got rid of the
sheet and contents of the sheet appearing to move independently as the
keyboard would appear or disappear.

As part of making that change, the sheet portion is broken out as its own
member variable. It has its own GeometryReader. The change also improves
readability a bit as the definition of |body| is now smaller.