Set backdrop color properly (#287)
This eliminates the unnecessary toggling of the BrowserViewController's alpha
value. Setting this to non-opaque was allowing the color of the underlying
window to show through. This is why we needed to ensure that the color of the
underlying window was set correctly to avoid having the backdrop for the
Tracking & Neeva menus appear lighter than it should.

With the change, the code gets a bit simpler. We can also revert the recent
change to BrowserTrayAnimators so the code is again as it was for Firefox.

Issue #286