Remove black flash when opening a new tab (#301)
Based on the comment, it appears that this code is part of the nightmode
implementation that we don't use. We'd probably have to revisit this change if
we do try to surface that feature to users.

Also, this fix only avoids the black flash when in light mode. In dark mode,
there is still a black flash, but that might be reasonable on dark mode. Not
specifying a background color for the WebView just means that the color of
the parent view will show through. Seems sensible.

Fixes #286