Allow user to dismiss add-to-spaces sheet by gesture (#395)
Previously, when the sheet was in its top most position, a drag down would
always move it to its half-height position. That's not great if the user
dragged the sheet down below half-height. So, this change lets a large drag do
just that.

Note, this could probably be smarter about calculating the height of the sheet
when in half-height mode and comparing the drag delta against that. I didn't
see an easy way to get access to that value though, so I went with a threshold
based approach.