Fix-up tinting of disabled toolbar buttons (#391)
Two changes:
- Disable the automatic tinting of disabled buttons
- Set the disabled button color specified in our mocks

Works properly in both light and dark modes.

Previously, the disabled tint color was in addition the default tinting
applied to images of a UIButton when disabled. That tinting didn't apply
to the text label of a UIButton. The add-to-spaces button is implemented
using a Nicon and so it is the text label portion.

What was happening here is that back/forward button disabled color looked
almost right because a fairly dark disabled tint color that we specify
was being combined with the extra tinting applied automatially by UIKit
for disabled buttons. But, it wasn't doing that for add-to-spaces. As a
result, add-to-spaces button did not appear to be changing color when
disabled. Actually, it was. It was just a very subtle dark gray.

Now, with automatic tinting disabled, we set an explicit disabled color
that applies to both the image and the text label of our toolbar buttons.

Fixes #383