Add dummy TabLocationView at the top of the Tab Tray (#422)
When tapped, opens a new tab and puts focus in the nav bar.

This works by creating a dummy instance of TabLocationView that just
serves to display the placeholder text and handle input. It could have
also worked to create this with a lighterweight control but it seemed
nice to reuse TabLocationView instead.

Other changes:
- Update background color of Tab Tray
- Remove dark-to-light transition of nav bar when selecting a tab.

In the future, we may want to improve the transition between real
TabLocationView and dummy TabLocationView. It can be a bit glitchy
in this implementation.

Issue #336