Speed up add-to-spaces by reusing previously fetched URLs (#428)
See discussion at https://github.com/neevaco/neeva/pull/41193 for the approach
on using the lastModifiedTs field to build and validate cached URL lists.

Fetching spaces data involves two queries: one to get the list of spaces and a
second to fetch data (URL list) for each of the spaces. The latter step makes
use of a new variant of the getSpace call to request data from many spaces in
one request.

The list of spaces is always fetched on refresh. That provides an up-to-date
view on lastModifiedTs values for each space. If the value of the lastModifiedTs
is unchanged for a given space, then the cached URL lists for that space are
considered up-to-date.

Further optimizations could include updating the lastModifiedTs on mutation.
This would avoid the need to update the URL list after adding an URL to spaces.

Issue #105