Set the DeviceType=ios-browser cookie as well (#444)
* Set the DeviceType=ios-browser cookie as well

Unfortunately when processing a HTTP server redirect, WebKit does not apply
changes we make to the WKWebView.customUserAgent property. It instead just
applies the default User-Agent string (or perhaps forwards whatever User-Agent
string was used for the redirected request). This means we cannot rely on
changing the User-Agent string to tell neeva.com that the request is from the
iOS browser.

An alternative to setting a custom User-Agent is to set the DeviceType cookie
with a value of "ios-browser". It may be possible to only do this, but this
change does not go that far. For now, both mechanisms are in place. I'd like to
run a few more tests before proposing that we stop setting a custom User-Agent
string.

A concern I have is that there may be a race condition between setting the
cookie and the network layer reading from the cookie store. Since the network
layer is in a different process, and indeed setCookie has an optional async
completion callback, it is hard to know if we are always going to win the race
here.

Fixes #441

* Actually, better to make the cookie persistent.

That way if in a cold start scenario we are first loading some
non-neeva page that happens to redirect to neeva, we will already
have the cookie set. Otherwise, we are betting on the cookie being
set in webView:decidePolicyFor always getting set before the HTTP
layer reads from the cookie jar. I'm not confident we will always
win that race.