Add NeevaFeatureFlags and set ClientOverride cookie (#522)
This change brings support for fetching client visible feature flags from Neeva.

Those are managed by the NeevaFeatureFlags class. Importantly, it implements
a stable view on the feature flags for the lifetime of the app. Updates to the flags
are written to defaults for use when the app is next launched.

A subset of Neeva feature flags need to be set by the browser on neeva.com so
that those features can be enabled in unison across the two. This is accomplished
with the ClientOverride cookie. That works like the FeatureFlagOverride cookie
except that it is settable by a non-neeva.co account.

See neevaco/neeva#41696 for more details on the change
that introduced the ClientOverride cookie and some of the rationale for the design.

This change also provides a way via debug settings to override flags locally. This is
useful for testing purposes. A very elementary UI is added to the existing feature
flags panel in settings. (Could use some UI love for sure.)

Issue #85