Use the same color for creating fallback letter favicons (#713)
* Use the same color for creating fallback letter favicons

From the comments, it looks like the intention is to make the background color for the letter part and the remaining part to be the same. But that hash is for sure not stable :). I am seeing too many varying colors in our suggestions. This always makes sure we are using one color for fallback

* Fix test

* Use named return variables