Only trust estimatedProgress when isLoading is true (#1226)
* Only trust estimatedProgress when isLoading is true

Unfortunately, when a request is cancelled, WebKit gets into a state
where it reports non-zero estimatedProgress and isLoading as false.
We just blindly trust estimatedProgress and use that to indicate if
the stop button should be visible, etc. What we should do instead is
look at both isLoading and estimatedProgress and only look at what
estimatedProgress says when isLoading is true.

Using combineLatest makes this an easy fix.

Fixes #1039

* Fix test failures.

Set alpha of progress bar to 0 when hiding. This way the a11y based
tests see the progress bar as nonexistent.

* Let progress done animation run to completion

And ensure reload button is in reload mode when not loading.

* Make sure progress is updated properly when switching tabs too