Clear login state when we land on neeva.com/signin (#1274)
* Clear login state when we land on neeva.com/signin

This way we avoid sending a stale httpd~login cookie to neeva.com. This may
help avoid an issue users are reporting with having trouble signing in after
having previously been signed in. One user noted that they had to explicitly
signout of the app through Settings in order to be able to sign in again.

Issue #269

* Move implementation to didFinish notification

This puts the code for cookie deletion next to the code where we set the
cookie in the first place.

Also reworks NeevaUserInfo so we can have all of the cookie juggling code
in one file.

Adds more information to the network logging to help with future debugging.

* Add test

* Refer to '/signin' using a named constant