Ensure the directory for the favicon cache exists (#1316)
Also do this lazily rather than at application startup time. The
additional I/O cost should be minimal, and this way we don't have
to worry about remembering to call a special setup function on the
FaviconFetcher at startup.

The existing code also worried about fxfavicon being a file rather
than a directory. Maybe that has to do with the way Firefox used
to work, but seems irrelevant for us.