Fix for cards getting stuck partially swiped (#1435)
SwiftUI DragGesture events unfortunately get interrupted by UIScrollView when
scrolling. See https://developer.apple.com/forums/thread/655465 where others
have reported this issue. The result is that `onEnded` is not reached, leaving
the card in a partially swiped state.

The workaround here is to simply avoid starting the swipe of the cards if the
drag gesture is more in the vertical direction than in the horizontal direction.
We have to care to allow such drag events once we have started swiping. This
seems to be okay as UIScrollView does not start scrolling in this case.

Fixes #1378