Native spaces part 2 (#1437)
Makes SpaceEntityData a BrowserPrimitive and adds CardDetails for it. Space is an AccessingManager for SpaceEntityData
Changes onSelect logic to allow selecting Spaces and SpaceEntities
Builds a DetailsView for any CardDetails that is also a ThumbnailModel for a list of CardDetails (which is Spaces and TabGroups right now)
DetailsView shows a list by default right now with only the thumbnail and the title of the entity.
Tweaks to the grid animation after topBar changes made us not have a predefined toolbar height. Now setting that in the model to make the GridPicker fade in and out properly with toolbar.
Added a toggle to show a grid.