Attempt to load sign-in / sign-up pages even if we are offline (#1446)
This fixes a regression. This code is using `NetworkReachability` incorrectly.
When first initialized, it does not immediately set `isOnline`. Instead that
gets set asynchronously. When used as part of a SwiftUI scene, this is fine.
When polled directly as in this code it does not work--it always just returns
`nil` after initialization.

Instead of trying to check if we are online, we should instead just load the
sign in / sign up pages. If they fail, the network error page will explain the
error and provide the user with an option to try again after they are online.
That's better than just not loading anything and dropping the user on the new
tab page.

Fixes #1445