Only dismiss the presentee if it was presented (#1458)
`Popover<Content>.ViewController` will automatically dismiss the presented view
controller when it disappears. However, it has a bug in that it will dismiss
whatever view controller happens to be presented. This change make sure that it
only dismisses a view controller that it was responsible for presenting.

This fixes the bug where we have observe the browser oscillating between showing
the IntroViewController and the BrowserViewController. BVC when it appears will
attempt to show the IntroViewController, and BVC when it disappears will trigger
the shield button to also disappear, which will attempt to close the shield
menu. The shield menu isn't showing, but since that uses `Popover<Content>`, it
will happily tear down the current view controller, i.e. IntroViewController.

That in turn shows the BVC, which causes it to show the IntroViewController again
because that is what it does when it appears (if the IntroViewController has not
yet been dismissed by the user).

Fixes #1455