Refactor code for setting up and using an overlay sheet (#1558)
This change introduces a showAsModalOverlaySheet method on BrowserViewController. This takes a SwiftUI View, and internally constructs a UIHostingController. This means each use of OverlaySheetView no longer needs its own UIHostingController.

This means we avoid duplication of code and eliminate some complexity in the process. This is also a step in the direction of having a single SwiftUI view hierarchy.

This PR converts AddToSpace, OpenInPage, Download and HTTPAuth dialogs. Overflow, Neeva Menu and Cheatsheet are left for a follow-up PR.

Issue #1409