Remove BVC dependency in SpacesModel (#1617)
The way it is built right now, it will be problematic to depend on BVC on SpacesModel. I would like to keep the recommended spaces model as a StateObject so that UI is self contained.

BVC is only needed for showing a "Item Added" toast inside the CardStrip in iPad after a drop which I imagine is not used at all today. So removing this dependency to add back later if needed.