Make swipe-to-close gesture less janky and scale better (#1641)
* Make swipe-to-close gesture less janky and scale better

The previous logic using predictedEndTranslation proved to be fairly
janky. The prediction is not always larger than the actual, etc.,
which makes sense.

A simpler approach of just expecting the displacement to be one card
size in length works better. This requires the gesture to be large
enough to avoid accidental swipes and also makes it work in a more
consistent way.

* Fix tests

- Switches from using swipeLeft() to tap()
- Switches EmptyCardGrid to be the background of CardsContainer

The first change is needed since the new thresholds for swiping
tabs to discard is no longer covered by the swipeLeft gesture.
A larger swipe is needed. The swipeLeft(velocity:) method does
not seem sufficient to account for this either. We should just
be able to tap the close button though!

The second change is needed to avoid a strange behavior of SwiftUI
in which the close button on the first opened tab would not be
hittable. Unclear why this happens, but it seems sensible to treat
this as the background view rather than a replacement view.

* Try to do the right thing for tab groups and native spaces

* Revise per review feedback

* Change definition of isCardGridEmpty