Make sure to call endBackgroundTask in all cases (#1669)
* Make sure to call endBackgroundTask in all cases

The apple documentation says we need to call endBackgroundTask from the
expirationHandler.

This change also "fixes" a blind call to Profile._shutdown in the AppDelegate.
This should probably also be wrapped with beginBackgroundTask just in case that
is respected by iOS at this point.

As part of that change, `getAppDelegate()` is introduced. With the addition of
that method `getAppDelegateProfile` is replaced w/ `getAppDelegate().profile`.

Also, this change makes sure to call `preserveTabs` on all of the `TabManager`
instances when the app is being terminated by the OS. This involved providing
a way to enumerate `TabManager` instances. (It is not possible to walk the
application's `connectedScenes` at this point.)

A WeakList<TabManager> is leveraged for that. There may be other cases in
the future when it is useful to be able to enumerate all tabs, and this makes
that possible. Seems better than keeping all `SceneDelegate` instances in a
weak list.

Issue #1652

* Cleanup profile accessor on AppDelegate