Introduce BackgroundFileWriter and use for saving tabs data (#1666)
* Introduce BackgroundFileWriter and use for saving tabs data

This utilizes the beginBackgroundTask API to give some protection
to our tabs saving code, which write data to a file on a background
thread.

Generalized the logic into a separate class (BackgroundFileWriter).
This class handles the details of writing to the same file multiple
times and scheduling that work to happen on a background thread.

This change also takes care to serialize the saved tabs array on
the main thread instead of from the background worker thread. That
should help protect against data corruption bugs where the main
thread may have been changing the saved tabs data structure while
the background thread was reading from it. Oops!

Issue #1652

* Add tests

* Factor out code to check file contents against expected value