Avoid populating models and views around recommended spaces for logged out users (#1931)
we are not using these spaces models, since we are relying on webui for the zero
query recommended spaces. so no need for the bloat and seems like this code has bit rotted
a bit. Removing the parts populating the view to be on the safe side about the public space
logged out experiment