Make sure card is in view when returning to the grid (#1975)
Without this change, if you tap on "show tabs" and scroll the grid
so that the selected tab is out of view, then when you hit "done"
and then "show tabs" again, the result will be that the tab is not
in view. Instead, we should scroll the card grid to make the selected
tab visible when opening the tab switcher.

The above is how Safari and Chrome behave, and I think it helps the
user not feel disoriented when they open a card grid and can't see
the selected tab.

Watch when the CardTransitionAnimator finishes and is hidden. Then
if the card grid is also hidden, call scrollTo. Also do this at
startup when the CardsContainer is created. Also do this whenever
the selected tab changes, again applying the constraint that the
card grid has to be hidden.