Simplify code (#2040)
No need to check if we have a login cookie with the callback from
updateLoginCookieFromWebKitCookieStore since that callback only
runs when there is a valid login cookie. (It's job is to copy the
login cookie from the WebKit cookie jar.)

We can also move the spaces refresh outside of the asyncAfter block
for the search bar prompt. They are okay to be unrelated.