Rework how we scroll the card grid to reveal the selected tab (#2054)
This time it is really fixed :)

Introduces a way to trigger scrollTo and then triggers that in the
appropriate places: when the card grid appears, when about to show
the card grid and when the card grid is hidden. The scrollTo animates
instantly, so it can happen at the start of the animation to show the
card grid without the user seeing any jumping.

Also rebuilds the ScrollView for the tabs whenever the devices switches
between portrait and landscape mode.

Changed updateCardSize to not get invoked when GridModel.isHidden
changes. This is something I noticed in passing and seemed good to
cleanup.

Fixes #2050