Add richEntities to Spaces (#2078)
* Add richEntities to Spaces

This adds all the necessary plumbing, pulls out info to put into a regular space entity detail,
builds a NavigationLink to display the details.

This is pretty functional with KIs around usability. Will follow up with ways to make the preview
a bit more prominent. And also we might want to touch up on the way the top label shows while viewing the WebView

* Remove the weird navigation bar and hide on click

Scroll works properly and we dont want the user to click and go to a different url here anyway
This hides the seam of using a webview by limiting functionality

* Add a label for Preview and swift format

* Remove the label for now

This doesnt change the url as desired

* Add the preview label back in

* Address comments