Add tech docs title and previews to spaces UI (#2117)
For tech docs (programming previews), the changes in native Spaces is pretty
light. We make sure the title is updated and there is a preview available

One thing this current PR doesnt address is tech docs with multiple sections. This
would require a way to open multiple previews which will need some UX help