Cache location of all loaded Cards to resolve more transition issues (#2134)
Fixes a bug in which switching tabs via the `createOrSwitchToTabs`
feature results in the `CardTransitionAnimator` picking the wrong
destination frame for the animation to show the tab tray.

This change makes it so that we aren't worried about keeping the
location of the selected Card up-to-date. Instead, we cache all
loaded Cards, and then just look for the location of the currently
selected Card.

This uses a bit more memory, but avoids race conditions. With this
fix, we can eliminate the `DispatchQueue.main.async` previously
added to `GridModel.hideWithAnimation` in an attempt to avoid those
race conditions.