Switch to showing web page when opening link from History (#2211)
* Switch to showing web page when opening link from History

This is for the "Overflow Menu in Card Grid" feature. When opening
the History Panel from the Card Grid, we should transition directly
to the web page when the user selects a link rather than showing
a blank card in the Card Grid.

This PR takes care to animate from the Card Grid when selecting an
existing tab. Otherwise, it skips the animation and just shows the
newly loading web page.

This adds a return value from createOrSwitchToTab{ForSpace} to let
the caller know if a tab was created or not. This isn't need for
the ForSpace variant, but it might be useful in the future, so
included it.

This also corrects an animation glitch that occured by calling
createOrSwitchToTab before the History Panel has closed. See
issue #2184 for more details on that. The fix for this glitch is
to delay calling createOrSwitchToTab in the completion callback
for dismissing the History Panel.

Issue #2184

* Use hideCardGrid