Fix crash by removing use of unowned (#2242)
It's possible for the `coordinator` to outlive the instance of
`TopBarHost`. Therefore we cannot safely use `[unowned self]`
here. There also doesn't look to be that much benefit to doing
so. Since the callback is a one-time callback just for the
animation, it is okay for it to retain a reference to `self`.

Fixes #2237