Call getKeyWindow when we are sure to be able to get it (#2246)
The existing code calls getKeyWindow(for: view) from within a
DispatchQueue.main.async block. When that code runs, it is
possible that the scene has already closed or the view has been
detached from the window, etc.

Getting the window from within viewDidLoad ensures that we will
be successful finding the window. Then we just hold onto the
window for use later in the async callback.

Fixes #2235