Fix bug with Pro Max phones showing 3 columns in card grid (#2257)
Since topToolbar is already a function of horizontalSizeClass,
checking horizontalSizeClass here would appear to be redundant.

But there's something more subtle going on. The topToolbar value
is passed as a parameter, and as it is coming from useEffect, it
is the value that topToolbar was and not necessarily is.

What I observed is that polling horizontalSizeClass from within
updateCardSize was returning .regular when topToolbar was set to
false. They are just temporarily out of sync during a rotation.

With this change, we simply avoid that problem, and we also
ensure that topToolbar and number of columns are kept in sync.

Fixes #1768