No need to copy preview cookie from keychain (#2283)
If the preview cookie is invalid, the server will generate a new one.
That happens via redirect.

Without this change, we can get into a situation with an endless
redirect loop. /search will see an invalid preview cookie and will
navigate the user to /preview, which will set a new preview cookie
and 302 redirect back to /search.

In the case of a 302 redirect, the browser does not notice the new
cookie in the cookie jar, and so when issuing the redirect request
it will still insert the old preview cookie.

We could also look at updating the keychain when a redirect occurs.
That may be wise to do, but a simpler solution is to just not copy
the preview cookie from the keychain into webkit's cookie jar.