Move card transition state updating to CardGrid (#2402)
This is an attempt to resolve the bug that we have sometimes seen
where the card transition does not complete, and the user is stuck
dealing with a zoomed tab that can just be moved up and down but
is otherwise unresponsive.

My theory is that the `CardTransitionModifier` is getting destroyed
too soon preventing the animation from completing and driving the
`runAfter` on `CardGrid`. However, I have not been able to fully
confirm this theory.

Moving all of the logic for updating `gridModel.isHidden` into the
`CardGrid` seems like a safer design at any rate, as it will be ok
with the `CardTransitionModifier` being removed prematurely. It
just means that we might observe a clipped transition animation,
but at least the browser should remain responsive.

This change also introduces a more readable `onCompletedAnimation`
in place of `runAfter` along with a specialization for `Bool`.

Fixes #2181